pip-requirements.txt
django==3.0.5
python-telegram-bot==12.6.1
sklearn==0.0
seaborn==0.10.1
matplotlib==3.2.1
numpy==1.18.3
djangorestframework==3.11.0
